removed some scary usize multiplications

+++ pumpkin-world/src/chunk/palette.rs
@@ -14,7 +14,7 @@ use super::format::{
 };
 
 /// 3d array indexed by y,z,x
-type AbstractCube<T, const DIM: usize> = [[[T; DIM]; DIM]; DIM];
+type AbstractCube<T, const DIM: u64> = [[[T; DIM]; DIM]; DIM];
 
 #[derive(Debug)]
 pub struct HeterogeneousPaletteData<V: Hash + Eq + Copy, const DIM: usize> {
@@ -116,7 +116,7 @@ impl<V: Hash + Eq + Copy + Default, const DIM: usize> PalettedContainer<V, DIM> 
                             debug_assert!((1 << bits_per_entry) > *key_index);
 
                             let packed_offset_index =
-                                (*key_index as u64) << (bits_per_entry as usize * index);
+                                (*key_index as u64) << (bits_per_entry as u64 * index as u64);
                             acc | packed_offset_index as i64
                         })
                     })
@@ -266,7 +266,7 @@ impl BiomePalette {
                             chunk.iter().enumerate().fold(0, |acc, (index, value)| {
                                 debug_assert!((1 << bits_per_entry) > *value);
                                 let packed_offset_index =
-                                    (*value as u64) << (bits_per_entry as usize * index);
+                                    (*value as u64) << (bits_per_entry as u64 * index as u64);
                                 acc | packed_offset_index as i64
                             })
                         })
@@ -348,7 +348,7 @@ impl BlockPalette {
                                 debug_assert!((1 << bits_per_entry) > *value);
 
                                 let packed_offset_index =
-                                    (*value as i64) << (bits_per_entry as usize * index);
+                                    (*value as i64) << (bits_per_entry as u64 * index as u64);
                                 acc | packed_offset_index
                             })
                         })
